Refactor: change package structure
